% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ci_mean.R
\name{get_ci_mean}
\alias{get_ci_mean}
\title{Calculate Confidence Intervals for the Mean}
\usage{
get_ci_mean(
  confidence = 0.95,
  sample_data,
  pop_sd = NULL,
  norm_dist = FALSE
)
}
\arguments{
\item{confidence}{Confidence level (default 0.95).}

\item{sample_data}{Numeric vector of sample data.}

\item{pop_sd}{Population standard deviation (default NULL).}

\item{pop_mean}{Population mean (not used currently, default NULL).}

\item{norm_dist}{Logical, whether to assume normal distribution (default FALSE).}
}
\value{
A data frame with mean, lower and upper bounds, and margin of error.
}
\description{
This function computes confidence intervals for the mean of a sample, depending on sample size and informations (if there are) about the population.
}
