% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_analysis.R
\name{text_analysis}
\alias{text_analysis}
\title{Text analysis}
\usage{
text_analysis(file_paths)
}
\arguments{
\item{file_paths}{A vector of file paths}
}
\value{
A list with word frequency and total sentiment score
}
\description{
This function reads text files, tokenizes them, counts word frequency, perform sentiment analysis and creates a word cloud.
}
